User Opt In
User is able to opt in or opt out of the memoryhole program

+++ memoryhole.py
@@ -5,13 +5,33 @@ import datetime
 import discord
 import os
 import time
+import json
 from redbot.core import commands
 
 DELETE_OLDER_THAN = datetime.timedelta(days=7)
+ids = {}
 
 class MemoryHole(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
+        try:
+            with open('memoryhole_ids.json') as f:
+                ids = json.load(f)
+        except FileNotFoundError:
+            print("Could not load memoryhole ids")
+
+    @commands.command(pass_context=True)
+    async def memoryhole(self, ctx):
+        """Toggles the user's inclusion in the memoryhole program"""
+        author = ctx.message.author.id
+        if author not in ids:
+            ids[author] = "1"
+            await ctx.send("You are now being memoryholed")
+        else:
+            del ids[author]
+            await ctx.send("You are no longer being memoryholed")
+        with open('memoryhole_ids.json', 'w+') as f:
+            json.dump(ids, f)
 
     async def loop(self):
         while self is self.bot.get_cog("MemoryHole"):
@@ -22,8 +42,9 @@ class MemoryHole(commands.Cog):
                     try:
                         async for message in channel.history(limit=None, before=datetime.datetime.utcnow() - DELETE_OLDER_THAN, after=None, oldest_first=True):
                             try:
-                                await message.delete()
-                                print("Deleted {}".format(message.clean_content))
+                                if message.author.id in ids:
+                                    await message.delete()
+                                    print("Deleted {}".format(message.clean_content))
                             except discord.errors.NotFound:
                                 pass
                             except discord.errors.Forbidden:
